pyproject.toml
[project]
name = "shining-light-backend"
version = "0.1.0"
description = "Shining Light School Management System"
authors = [
    {name = "Urkel",email = "abmichael109@gmail.com"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "django-filter (>=25.2,<26.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "dj-rest-auth (>=7.0.1,<8.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=25.9.0,<26.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "isort (>=6.1.0,<7.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
    "faker (>=37.11.0,<38.0.0)"
]
